good

+++ menu_data.php
@@ -5,11 +5,24 @@
 
     $CFG = include_once("../common/include/incConfig.php");
 
+    require_once("../common/include/incUser.php");
+    require_once("../common/include/incAuth.php");
     require_once("../common/include/incUtil.php");
     require_once("../common/include/incDB.php");
     require_once("../common/include/incSec.php");
 
     //ServerViewTxt("N","N","Y","Y");
+    $resToken = uniqid();
+
+    $log = getLogger(
+        array(
+        "LIST_NM"=>"log_CG"
+        , "PGM_ID"=>"menu_data"
+        , "REQTOKEN" => $reqToken
+        , "RESTOKEN" => $resToken
+        , "LOG_LEVEL" => Monolog\Logger::INFO
+        )
+    );
 
     $db=getDbConn($CFG["CFG_DB"]["CGCORE"]);
 
@@ -34,13 +47,18 @@ alog("---------------GRP PGM ---------------------START");
           ";
 
     $REQ = null;
-    $stmt = make_stmt($db,$sql, $to_coltype, $REQ);
+    $stmt = makeStmt($db,$sql, $to_coltype, $REQ);
     if(!$stmt)   JsonMsg("500","108","stmt 생성 실패" . $db->errno . " -> " . $db->error);
     //var_dump( make_grid_read_array($stmt) );
+    if(!$stmt->execute())JsonMsg("500","100","stmt 실행 실패" . $db->errno . " -> " . $db->error);
+    
 
-    $tResultArray = make_grid_read_array($stmt);
-    $stmt->close();
-    foreach($tResultArray->RTN_DATA->data as $tMap) {
+    //$tArr =  getStmtArray($stmt);
+
+    //$tResultArray = make_grid_read_array($stmt);
+    //$stmt->close();
+    //echo "<br>리턴 결과 row수 : " . getRowCount($stmt);    
+    while( $tMap = $stmt->fetch(PDO::FETCH_ASSOC) ) {
         ?>
         <item   text="<?=$tMap["PJTNM"]?>" id="<?=$tMap["PJTID"]?>" open="1">
         <?php
@@ -55,29 +73,75 @@ alog("---------------GRP PGM ---------------------START");
 
             $dbPjt=getDbConn($CFG["CFG_DB"][$tMap["DSNM"]]);
 
-        
+            /*
+            if(function_exists($dbPjt->getAttribute)){
+                echo "pdo_mysql";
+            }else{
+                echo "mysqli";
+            }
+            */
 
             $REQ["PJTSEQ"] = $tMap["PJTSEQ"];
+            //echo "<BR>DSNM : " .  $tMap["DSNM"] ;
+            //echo "<BR>PJTSEQ : " .  $REQ["PJTSEQ"] ;
             $to_coltype = "i";
-            $sql = " select PGMSEQ,PGMID,PGMNM,VIEWURL from CG_PGMINFO where PJTSEQ = #PJTSEQ# ";
+            $sql = " select PGMSEQ,PGMID,PGMNM,VIEWURL from CG_PGMINFO where PJTSEQ = #{PJTSEQ} ";
     
-            $stmt = make_stmt($dbPjt,$sql, $to_coltype, $REQ);
+            $stmt2 = makeStmt($dbPjt,$sql, $to_coltype, $REQ);
     
     
-            if(!$stmt)   JsonMsg("500","109","stmt 생성 실패" . $dbPjt->errno . " -> " . $dbPjt->error);
+            if(!$stmt2)   JsonMsg("500","109","stmt 생성 실패" . $dbPjt->errno . " -> " . $dbPjt->error);
+            if(function_exists($dbPjt->getAttribute)){
+                //pdo
+                try{
+                    //echo "실행결과 : " . $stmt2->execute(array($REQ["PJTSEQ"]));
+
+                    if(!$stmt2->execute(array($REQ["PJTSEQ"]))){
+                        $arr = $stmt2->errorInfo();
+                        print_r($arr);
+                        //JsonMsg("500","100","stmt 실행 실패" . $dbPjt->errno . " -> " . $dbPjt->error);
+                    }
+                } catch(PDOException $e) {
+                    echo "PDOException : " . $e->getMessage();
+                }
+                
+            }else{
+                
+                //mysqli
+                if(!$stmt2->execute())JsonMsg("500","100","stmt 실행 실패" . $dbPjt->errno . " -> " . $dbPjt->error);
+                
+            }
+
+
             //var_dump( make_grid_read_array($stmt) );
             //echo "<br>111";
     
-            $tPgmArray = make_grid_read_array($stmt);
-            $stmt->close();
+            //$tPgmArray = make_grid_read_array($stmt);
+            //$stmt->close();
             $subItemCnt = 0;
-            foreach($tPgmArray->RTN_DATA->data as $tMap2) {
-                echo '      <item text="' . $tMap2["PGMNM"] . '" id="' . $tMap2["PGMSEQ"] . "^" . $CFG["CFG_MAKE_URL"] . "/c.g/" . $tMap["PJTID"] . "/" . $tMap2["VIEWURL"] . '"></item>' . PHP_EOL;
-                $subItemCnt++;
+
+            if(function_exists($dbPjt->getAttribute)){
+                //pdo
+                //echo "444";                
+                while( $tMap2 = $stmt2->fetch(PDO::FETCH_ASSOC) ) {
+                    echo '      <item text="' . $tMap2["PGMNM"] . '" id="' . $tMap2["PGMSEQ"] . "^" . $CFG["CFG_MAKE_URL"] . "/c.g/" . $tMap["PJTID"] . "/" . $tMap2["VIEWURL"] . '"></item>' . PHP_EOL;
+                    $subItemCnt++;
+                }
+            }else{
+                //mysqli
+                //echo "333";
+                $result = $stmt2->get_result();
+                //echo "<br>리턴 결과 row수 : " . getRowCount($result);
+                while ($tMap2 = $result->fetch_array(MYSQLI_ASSOC)){
+
+                    echo '      <item text="' . $tMap2["PGMNM"] . '" id="' . $tMap2["PGMSEQ"] . "^" . $CFG["CFG_MAKE_URL"] . "/c.g/" . $tMap["PJTID"] . "/" . $tMap2["VIEWURL"] . '"></item>' . PHP_EOL;
+                    $subItemCnt++;
+                }
             }
-            //echo "<br>222";        
-            $dbPjt->close();
-            unset($dbPjt);
+
+
+            closeStmt($stmt2);
+            closeDb($dbPjt);
         }
 
         if($subItemCnt == 0){
@@ -91,7 +155,9 @@ alog("---------------GRP PGM ---------------------START");
                 //echo "<br>333";
     }
     //echo "<br>444";
-    $db->close();
+    closeStmt($stmt);
+    closeDb($db);
+    
 ?>        
 
 </tree>
